--- apps/mech/common-util/Login/index.jsx
@@ -1,2 +1,0 @@
-export { SUPPORTED_CHAINS } from './config';
-export { LoginV2 } from './LoginV2';
